pkg/pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "versioningit"]
build-backend = "setuptools.build_meta"


# ----------------------------------------- setuptools -------------------------------------------
[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true


# ----------------------------------------- Project Metadata -------------------------------------
#
[project]
version = "0.0.0.dev61"
name = "PySerials"
dependencies = [
    "jsonschema >= 4.21.0, < 5",
    "referencing >= 0.35.1",
    "jsonpath-ng >= 1.6.1, < 2",
    "ruamel.yaml >= 0.18", # https://yaml.readthedocs.io/en/stable/
    "ruamel.yaml.string >= 0.1.1, < 1",
    "tomlkit >= 0.11.8, < 0.14", # https://tomlkit.readthedocs.io/en/stable/,
    "MDit == 0.0.0.dev58",
    "ExceptionMan == 0.0.0.dev58",
    "ProtocolMan == 0.0.0.dev2",
]
requires-python = ">=3.10"
